Auth access

+++ part3/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Reviews operations')
 
@@ -18,30 +19,36 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """
         Register a new review
         """
         review_data = api.payload
+        current_user = get_jwt_identity()
+
+        review_data["user_id"] = current_user["id"]
 
         existing_user = facade.get_user(review_data["user_id"])
         if not existing_user:
             return {'error': 'User not found'}, 404
 
-        review_data["user"] = existing_user
-
         existing_place = facade.get_place(review_data["place_id"])
         if not existing_place:
             return {'error': 'Place not found'}, 404
 
-        review_data["place"] = existing_place
+        if existing_place.owner_id == current_user["id"]:
+            return {'error': 'You cannot review your own place'}, 400
+
+        for review in existing_place.reviews:
+            if review.user_id == current_user["id"]:
+                return {'error': 'You have already reviewed this place'}, 400
 
         try:
             new_review = facade.create_review(review_data)
         except ValueError:
             return {'error': 'Invalid input data'}, 400
         else:
-            existing_place.add_review(new_review)
             return {
                 'id': new_review.id,
                 'title': new_review.title,
@@ -94,14 +101,22 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """
         Update a review's information
         """
         review_data = api.payload
+        current_user = get_jwt_identity()
+
+        review_data["user_id"] = current_user["id"]
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+
+        if review.user_id != current_user["id"] and not current_user["is_admin"]:
+            return {'error': 'Unauthorized action'}, 403
+
         updated_review = facade.update_review(review_id, review_data)
         if not updated_review:
             return {'error': 'Invalid input data'}, 400
@@ -109,12 +124,19 @@ class ReviewResource(Resource):
 
     @api.response(204, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """
         Delete a review
         """
+        current_user = get_jwt_identity()
+
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+
+        if review.user_id != current_user["id"] and not current_user["is_admin"]:
+            return {'error': 'Unauthorized action'}, 403
+
         facade.delete_review(review_id)
         return {'message': 'Review deleted successfully'}, 204
